refactor: extract affinity.
